code/ci360-audience-management-utility/requirements.txt
requests
pyjwt
pandas
configparser